Backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.13.2
pydantic-settings==2.14.0
python-jose[cryptography]==3.5.0
pwdlib[email-validator]==0.3.0
email-validator==2.3.0
python-multipart==0.0.26